move common assertions and test parameter providers into the GitIntegrationTest
